Monster class: spell parsing basics
